Precompute DFS of all nodes
mergiraf currenly spends a lot of time in the DFS building code. On
MacOS when processing a PNPM lockfile, the majority of time (of a 10s
sample) was spent in the allocator on the `isomorphic_to` stack (where
the allocations had to be from DFS creation). On GNU/Linux it's less
significant, but still makes a difference.

We observe that the (prefix) DFS of a node is a subslice of the DFS of
its parent node. This means that all we need to do is compute the DFS of
the root node once, and then let the children store slices into that to
effectively precompute it for the entire tree.

By doing this, we significantly speed up the process. This benchmark was
done on a truncated (~4k loc) PNPM YAML lockfile with a one-line
change/conflict.

```
Benchmark 1: mergiraf-after merge --git .merge_file_6vIArC .merge_file_KYG0x3 .merge_file_9waipA -s fb4f872 -x HEAD -y branch1 -p a.yaml
  Time (mean ± σ):      5.648 s ±  0.074 s    [User: 5.616 s, System: 0.057 s]
  Range (min … max):    5.580 s …  5.770 s    5 runs

Benchmark 2: mergiraf-before merge --git .merge_file_6vIArC .merge_file_KYG0x3 .merge_file_9waipA -s fb4f872 -x HEAD -y branch1 -p a.yaml
  Time (mean ± σ):      6.877 s ±  0.133 s    [User: 6.857 s, System: 0.053 s]
  Range (min … max):    6.699 s …  7.016 s    5 runs

Summary
  mergiraf-after merge --git .merge_file_6vIArC .merge_file_KYG0x3 .merge_file_9waipA -s fb4f872 -x HEAD -y branch1 -p a.yaml ran
    1.22 ± 0.03 times faster than mergiraf-before merge --git .merge_file_6vIArC .merge_file_KYG0x3 .merge_file_9waipA -s fb4f872 -x HEAD -y branch1 -p a.yaml
```