don't add `parsed_merge` to `merges`
the current pipeline looks like this:
1. `parsed_merge` is created as a result of `resolve_merge`
2. it gets added to `merges` (this makes the later check for empty
`merges` superfluous) as a `MergeResult` with `method="original"` (a
method that doesn't exist anywhere else in the codebase)
3. if it ends up being the only element in `merges`, it gets chosen as
`best_merge` and returned
4. in `mergiraf.rs`, `best_merge` is explicitly checked to _not_ have
`method="original"` -- as such, `parsed_merge` is never even returned.
Additionally, the error code from `Ok(parsed_merge)` and `Err(_)` is the
exact same

All in all, the addition of `parsed_merge` to `merges` doesn't have any
consequences and might as well not happen in the first place.

This changes the step 3 -- now, the empty `merges` error is returned,
and handled in `mergiraf.rs`