refactor(SUPPORTED_LANGUAGES): extract typescript settings
Due to the fact that `tree-sitter-typescript` has distinct grammars for
standard TypeScript and TSX, we'll need two separate language profiles.
The bulk of their configuration will be the same, however.

In preparation for supporting TSX, extract the language profile settings
that will be common to both standard TypeScript and TSX.