perf(normalize_to_lf): take and return an owned value in case no change is needed
The motivation from this came from #141, where we leak `contents` after
normalizing them, and as such would like to have an ownership of the
value. As it is now though, `normalize_to_lf` only takes a borrowed
`&str` and thus cannot take advantage of an owned value being available.

This commit changes that by effectively changing the argument's type to
a `Cow<str>` instead. In particular, `Into<Cow<str>>` was used to avoid
having to prepend `Cow::from` at every call site