refactor(DisplaySettings): use `Option`s for default fields
Previously, we needed to specify the default values in multiple places:
- the two `default{,_compact}` constructors
- in the binary, as `default_*_name`
- also in the binary, as the default values for flags

This commit removes all those and uses `None`s in there, instead adding
methods called `DisplaySettings::<field_name>_or_default` that return
`Some` values if they exist, or the _actual_ defaults in case of `None`.

There was another approach that came to mind, but was seen as too
cumbersome: make a newtype for each such field, and implement `Default`
for them -- this'd requier too much boilerplate and could require
exposing the newtypes in a lot of places (most importantly, in the
binary).

Ideally, of course, we'd use default field values, but those aren't
stabilzed yet.[^1]

This also gets rid of the TODO in `mergiraf solve`, which was addressed
back in #74

[^1]: https://github.com/rust-lang/rust/issues/132162