fix(highlight_duplicate_signatures): don't remove separators if no conflict was built
`conflict_add_separator` and the surrounding comments suggest that
what we're expecting here is to build a conflict between the duplicated signatures.
Because of that, when we're creating `conflict_add_separator` and see `add_separator=AtEnd`,
we command to remove the next separator, since we plan to have it inside the conflict sides
(so that when the conflict is resolved, we don't end up with duplicated separators).

But `merge_same_sigs` has another, happier path, where it realizes that the 2 nodes
with duplicated signatures are actually isomorphic. In that case, it just resolves the conflict.

And herein lies the problem: the fact that the conflict didn't get constructed after all
isn't taken into account by the caller -- it'll still remove the next separator
(and will have removed the previous separator, if we had `add_separator=AtBeginning`) all the same.

The reason why this wasn't caught by any of the tests before is that until now,
whenever we arrived at this happy path, `add_separator` has been `OnlyInside` --
so no erroneous separator removal happened. But that's seemingly because
`OnlyInside` is just what we have most of the time.

So the fix for this is not that complicated. We:
- don't yet remove the separators when constructing `conflict_add_separator`
- return from `merge_same_sigs` whether the happiest path was taken
- only remove the separators if it wasn't

Fixes #208