fix: restore the comment deleted by rust-analyzer
I first thought this is some bad merge by Mergiraf, but it was actually
innocent. The real reason was that when I rebased #240 onto #238, the
resulting file didn't import `languages`, and so I fixed the error by
running the code action without thinking about it. But it turns out that
when adding the import, rust-analyzer normalized the import set,
printing it all on one line and removing the comment in the process.
This then slipped through during the review of #240.

The comment also somehow affects the grouping of multiple imports on one
line, so there's that as well.

We are not the only ones struggling with the whitespace, are we.