perf(ParsedMerge::parse): only compile the regex once
This function is called from multiple places during the runtime of the
program -- in particular, each time a node requires line-based fallbackn
that case. The regex recompilations do add up.

This wraps the regexes in `OnceCell`s, so that they cab be reused across
calls.

As can be seen by the change to the test, this does mean that if
`ParsedMerge::parse` is called multiple times in a row[^1] with
different `conflict_marker_size`s, then only the first one will be
respected. This is very unlikely to be a footgun in practice though,
since `conflict_marker_size` never changes during any given run of
Mergiraf.

[^1]: parallel calls would probably cause a race condition and make the
test flaky