doc: explain why we pull trailing whitespace into conflict in compact mode
I was wondering why this is done and almost opened an issue. The problem
was worsened by the fact that the example provided in the corresponding
test wasn't making the motivation obvious either.

Now that I got it, I wrote the motivation down and changed the test to
use a "real-life" example (from an integration test)