Revert "trailing_whitespace: call `AstNode::trailing_whitespace` earlier"
That commit actually introduced a change into method logic. Before, we
first tried to get nodes at the three revisions, and proceeded based on
whether all of those were `Some`. But after the commit, we try to get
the nodes _and_ get their `trailing_whitespace`s, and match on the
success of _that_. I'm not sure if this new behaviour is still okay, but
I'd rather err on the safe side.

This reverts commit 25771fdcf60a6aa30d5801609bb0fc463c51780e, part of
#300