Improve error reporting when working copy is not in a conflict state
When trying to resolve a conflict with `mergiraf solve` in a working
copy which is not in a conflict state, mergiraf reports

```
error while retrieving Base revision for <path>:
git checkout-index: <path> does not exist at stage 1
```

This actually comes directly from `git checkout-index` but is
confusing as it's not clear what the issue could be. The underlying
issue is that the index entry is not in a conflict state, and so there
is no stage git can recover.

As `checkout-index` is a safeguard, just don't show its stderr and
display a mild warning that mergiraf could do better with the right
setup.