Call checkout-index just once per file
`extract_revision(_from_git)?` is always called 3 times (either
directly or indirectly via `extract_revision`, doesn't really make a
difference), one time to get the base version of a file, one time to
get the left, and one time to get the right.

However this is a lot of unnecessary git interactions given Git can do
that on its own by calling `checkout-index --stage=all`. It's a bit
more complex to parse but not exactly heinous.