chore: Move mgf_dev to a separate package
If mgf_dev is in the same package as the primary mergiraf binary,
`cargo install` will install both by default. When using cargo-binstall,
this will fail the installation, because the mgf_dev binary is not
included in our provided release packages.

With the binaries separated into distinct packages, cargo and related
tools will default to only the primary mergiraf binary for running or
installing.