fix: improve detection of Jujutsu usage
We try to detect if the command `mergiraf solve` is being run in a
Jujutsu repository, which can lead to problems.

The heuristic to detect Jujutsu repositories was to execute `jj root`
and check the exit code. However, that lead to false positives where
users had a completely different program installed that was also called
`jj` and also happened to exit with 0 when called with the argument
`root`.

The improved heuristic also checks if a `.jj` entry exists in the
directory returned by `jj root`.

The specific case of the reported false positive [^1] was a conflict
with a JSON stream editor called jj [^2]. It doesn't write anything to
stdout, so the new heuristic correctly accounts for that case.

[^1]: https://codeberg.org/mergiraf/mergiraf/issues/511
[^2]: https://github.com/tidwall/jj